requirements.txt
pandas==1.5.3
numpy==1.24.2
requests==2.28.1
tqdm==4.64.1
streamlit==1.18.1
plotly==5.14.1
pyyaml==6.0
pytest==7.2.1
